nacc-common/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "nacc-common"
version = "3.1.0"
description = "NACC Data Platform utilities"
license-files = ["LICENSE"]
authors = [{ name = "NACC", email = "nacchelp@uw.edu" }]
readme = "README.md"
dependencies = ["flywheel-sdk>=20.0.0", "python-dateutil>=2.5.3,<3", "pydantic>=2.5.2,<3", "typing_extensions>=4.0.0"]
requires-python = ">=3.10,<3.13"

[tool.setuptools.package-dir]
nacc_common = "src/python/nacc_common"
